api/requirements.txt
pymongo
python-dotenv
librosa
numpy
Flask